broke up the Dashboard section up into smaller components
