+++ client/src/components/BudgetCardList/BudgetCardList.styles.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  width: 100%;
+  max-width: 800px;
+  padding: 1em;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  gap: 2em;
+  align-items: center;
+`;
