+++ client/src/components/index.tsx
@@ -4,3 +4,5 @@ export { default as TransactionItemAdder } from "./TransactionItemAdder/Transact
 export { default as TransactionItem } from "./TransactionItem/TransactionItem";
 export { default as TransactionItemsList } from "./TransactionItemsList/TransactionItemsList";
 export { default as TransactionItemEditor } from "./TransactionItemEditor/TransactionItemEditor";
+export { default as BudgetCardList } from "./BudgetCardList/BudgetCardList";
+export { default as BudgetCard } from "./BudgetCard/BudgetCard";
