+++ client/src/dummydata.ts
@@ -1,23 +1,26 @@
 export const BudgetData = [
   {
+    _id: "1",
     month: "February",
     year: "2022",
-    budget: 3200,
-    spent: 3300,
+    total: 3200,
+    currentAmount: 3300,
     active: false,
   },
   {
+    _id: "2",
     month: "March",
     year: "2022",
-    budget: 3200,
-    spent: 1784,
+    total: 3200,
+    currentAmount: 1784,
     active: false,
   },
   {
+    _id: "3",
     month: "April",
     year: "2022",
-    budget: 3200,
-    spent: 500,
+    total: 3200,
+    currentAmount: 500,
     active: true,
   },
 ];
